Menus now work
Able to recruit ships and upgrade planets

subtracts the cost for the resources. All we have left to do is Add the
random events, and the effects of the buildings

and finally the increase of resources over time and I think we should be
good

+++ runPlanet/LargePlanet.pde
@@ -5,7 +5,7 @@
 
 public class LargePlanet extends Planet
 {  
-  LargePlanet(int xp, int yp, boolean a, int r, int s)
+  LargePlanet(float xp, float yp, boolean a, int r, int s)
   {
     super(xp,yp,a,r,s);
     size=size*5;
