+++ runPlanet/MedPlanet.pde
@@ -4,7 +4,7 @@
 
 public class MedPlanet extends Planet
 {  
-  MedPlanet(int xp, int yp, boolean a, int r, int s)
+  MedPlanet(float xp, float yp, boolean a, int r, int s)
   {
     super(xp,yp,a,r,s);
     size=size*3;
